[DDW-864] Implement Transaction details page design - adding Transaction summary component
