+++ source/features/transactions/components/TransactionInfo.scss
@@ -1,32 +1,116 @@
 .transactionInfoContainer {
   margin: 61.5px 0;
+  width: 872px;
 
   .header {
     margin-bottom: 41px;
   }
 
-  .content {
-    display: flex;
-    justify-content: space-between;
+  .transactionInfoRow {
+    margin-bottom: 35px;
 
-    .infoPanel {
-      color: $solid-text;
-      opacity: 0.7;
+    .transactionInfoRowContainer {
+      display: flex;
+      font-size: 12px;
+      justify-content: space-between;
+      letter-spacing: 0.4px;
+      line-height: 1.33;
 
-      .infoRow {
-        display: flex;
-        font-size: 12px;
-        letter-spacing: 0.6px;
-        line-height: 1.33;
-        margin-bottom: 10px;
+      .addresses {
+        width: 606px;
 
-        .infoLabel {
-          font-weight: 600;
-          width: 130px;
+        .infoRow {
+          align-items: center;
+          display: flex;
+          justify-content: space-between;
+          margin-bottom: 10px;
+
+          &:last-child {
+            margin-bottom: 0;
+          }
+
+          .currentAddress {
+            color: $solid-text;
+            opacity: 0.7;
+          }
+
+          .senders,
+          .receivers {
+            display: flex;
+            flex-direction: column;
+            width: calc(50% - 27px);
+          }
+
+          .senders {
+            align-items: flex-start;
+          }
+
+          .receivers {
+            align-items: flex-end;
+          }
+
+          .id {
+            color: $primary-highlight-color;
+          }
+
+          .sender,
+          .receiver {
+            color: $primary-highlight-color;
+            margin-bottom: 5px;
+
+            &:last-child {
+              margin-bottom: 0;
+            }
+          }
+
+          .transferredAt {
+            color: $solid-text;
+            opacity: 0.7;
+          }
+
+          .nextIcon {
+            display: flex;
+            justify-content: center;
+
+            svg {
+              height: 10px;
+              width: 14px;
+              object-fit: contain;
+            }
+          }
         }
+      }
 
-        .infoValue {
-          font-weight: 300;
+      .amounts {
+        flex: 1;
+
+        .infoRow {
+          align-items: flex-end;
+          display: flex;
+          flex-direction: column;
+          margin-bottom: 5px;
+
+          &:last-child {
+            margin-bottom: 0;
+          }
+
+          .totalAmount {
+            background-color: $highlighted-dark-area-bg;
+            border-radius: 1px;
+            color: $solid-text;
+            opacity: 0.7;
+            padding: 5px 10px;
+          }
+
+          .amount {
+            color: $solid-text;
+            margin-bottom: 5px;
+            opacity: 0.7;
+
+            &:last-child {
+              margin-bottom: 0;
+            }
+          }
         }
       }
     }
