--- source/features/transactions/components/TransactionList.scss
@@ -1,12 +1,0 @@
-.transactionListContainer {
-  margin: 61.5px 0;
-  width: 872px;
-
-  .header {
-    margin-bottom: 41px;
-  }
-
-  .transactionListRow {
-    margin-bottom: 35px;
-  }
-}
