--- source/features/transactions/components/TransactionList.tsx
@@ -1,25 +1,0 @@
-import { observer } from 'mobx-react-lite';
-import DividerWithTitle from '../../widgets/divider-with-title/components/DividerWithTitle';
-import styles from './TransactionList.scss';
-import TransactionListRow, {
-  ITransactionListRowProps,
-} from './TransactionListRow';
-
-export interface ITransactionListProps {
-  items: Array<ITransactionListRowProps>;
-}
-
-const TransactionList = (props: ITransactionListProps) => (
-  <div className={styles.transactionListContainer}>
-    <div className={styles.header}>
-      <DividerWithTitle title="Transactions" />
-    </div>
-    {props.items.map((item, index) => (
-      <div key={`transaction_${index}`} className={styles.transactionListRow}>
-        <TransactionListRow {...item} />
-      </div>
-    ))}
-  </div>
-);
-
-export default observer(TransactionList);
