+++ source/features/transactions/components/TransactionSummary.scss
@@ -1,0 +1,40 @@
+.transactionSummaryContainer {
+  margin: 61.5px 0;
+
+  .header {
+    margin-bottom: 41px;
+  }
+
+  .content {
+    display: flex;
+    justify-content: space-between;
+
+    .infoPanel {
+      color: $solid-text;
+      opacity: 0.7;
+
+      .infoRow {
+        display: flex;
+        font-size: 12px;
+        letter-spacing: 0.6px;
+        line-height: 1.33;
+        margin-bottom: 10px;
+
+        .infoLabel {
+          font-family: 'Proxima Nova Semibold', sans-serif;
+          font-weight: 600;
+          width: 130px;
+        }
+
+        .infoValue {
+          font-family: 'Proxima Nova L', sans-serif;
+          font-weight: 300;
+
+          span {
+            color: $primary-highlight-color;
+          }
+        }
+      }
+    }
+  }
+}
