+++ source/features/widgets/divider-with-title/components/DividerWithTitle.scss
@@ -7,9 +7,11 @@
     opacity: 0.5;
     width: 100%;
   }
+
   .title {
     background-color: $primary-bg;
     color: $solid-text;
+    font-family: 'Proxima Nova Semibold', sans-serif;
     font-size: 12px;
     font-weight: 600;
     left: 50%;
